Update Diagnostic category and ID ranges to reflect currently used IDs
The category to ID ranges in this file currently reflect the **reserved** ID ranges for the category. This change updates the CA ranges to now reflect the **currently used** ID ranges for the category. This makes it easier to identify the next available ID for a specific CA category when implementing a new rule. There will be an additional step of updating the range end in this text file when implementing a new rule, but that seems a good tradeoff versus spending additional time figuring out the correct next available ID to use. I will send a separate PR to add a code fix to automatically update this text file.